+++ coaster/views/__init__.py
@@ -10,3 +10,4 @@ Coaster provides classes, functions and decorators for common scenarios in view 
 from __future__ import absolute_import
 from .misc import *  # NOQA
 from .decorators import *  # NOQA
+from .classview import *  # NOQA